Cleaned headers + Move static function outside of class + Add comments.
